requirements.txt
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
git-filter-repo==2.47.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonify==0.5
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
oauthlib==3.2.2
openslide-bin==4.0.0.6
openslide-python==1.4.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
